Math related function are now in tml::Math namespace Instead of tml::Util.

+++ src/Interface/Components/Progressbar.cpp
@@ -17,10 +17,10 @@ namespace tml
         {
             DEFAULT_TEXT->SetString(std::to_string(ui32(m_value * 100)) + "%");
             DEFAULT_TEXT->SetSize(m_size.y / 2);
-            DEFAULT_TEXT->SetPosition({m_pos.x + (m_size.x / 2) - (DEFAULT_TEXT->GetDimensions().x / 2), m_pos.y + m_size.y / 4});
+            DEFAULT_TEXT->SetPosition({m_pos.x + (m_size.x / 2) - (DEFAULT_TEXT->GetDimensions().x / 2), static_cast<float>(m_pos.y + m_size.y / 4)});
 
             Renderer::DrawRect(m_pos, m_size, m_pColor);
-            Renderer::DrawRect(m_pos, {m_size.x * m_value, m_size.y}, m_activeColor);
+            Renderer::DrawRect(m_pos, {m_size.x * m_value, static_cast<float>(m_size.y)}, m_activeColor);
             Renderer::DrawGrid(m_pos, m_size, 1, 1, m_sColor);
             Renderer::Draw(*DEFAULT_TEXT);
         }
